requirements.txt
requests
beautifulsoup4
pyppeteer
retrying
python-dotenv
